potree-project-computengine/potreeproject/requirements.txt
appdirs==1.4.3
asgiref==3.2.2
asn1crypto==0.24.0
attrs==19.1.0
black==19.3b0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Django==3.0a1
django-rest-knox==4.1.0
djangorestframework==3.10.3
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.0
idna==2.8
mysqlclient==1.4.5
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2019.2
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.3
